fix(domain,icons,txn): quick correctness — from_hex panic, oracle parity, consent, overflow (audit F6/F7/F11)
fingerprint::from_hex indexes BYTES (a 64-byte string with a multi-byte char panicked slicing mid-codepoint). item::can_style now requires !requires_explicit_consent. background.rs uses FLOAT division to match the frozen JS oracle (integer 256/3=85 vs 85.333 diverged on an 85px board). txn::id + ledger::store use checked_add (from_journal/next_version Err on exhaustion; next_id panics rather than wrapping to a REUSED id).

+++ crates/dm-icon-core/src/analysis/background.rs
@@ -29,7 +29,10 @@ fn try_shape_background(c: &Raster) -> Option<Rgba> {
         return None;
     }
     let min_dim = bounds_w(bounds).min(bounds_h(bounds));
-    if min_dim < c.width / 3 {
+    // FLOAT division to match the frozen oracle (audit F7 / ADR-0019): integer `c.width / 3` moves
+    // the boundary (e.g. 256/3 = 85 in Rust vs 85.333 in JS), so an 85px board on a 256px canvas
+    // would diverge — Rust proceeding while the oracle returns none.
+    if (min_dim as f64) < c.width as f64 / 3.0 {
         return None;
     }
     // Owner law ②: only corner-symmetric silhouettes own a board.
